Move some functions from htlc_tests to db_fixture
